Moved properties to matlib folder. Removed module properties. Moved MaterialXProperties to utils.py. Log fixes. Code cleanup.
