+++ materialx/matlib/__init__.py
@@ -1,8 +1,17 @@
+# SPDX-License-Identifier: GPL-2.0-or-later
+# Copyright 2022, AMD
+
 import bpy
 
-from . import ui
+from . import ui, properties
 
 
+register_properties, unregister_properties = bpy.utils.register_classes_factory(
+    [
+        properties.MatlibProperties,
+        properties.WindowManagerProperties,
+    ]
+)
 register_ui, unregister_ui = bpy.utils.register_classes_factory(
     [
         ui.MATLIB_PT_matlib,
@@ -16,8 +25,10 @@ register_ui, unregister_ui = bpy.utils.register_classes_factory(
 
 
 def register():
+    register_properties()
     register_ui()
 
 
 def unregister():
     unregister_ui()
+    unregister_properties()
