+++ materialx/matlib/manager.py
@@ -1,17 +1,6 @@
-#**********************************************************************
-# Copyright 2020 Advanced Micro Devices, Inc
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#********************************************************************
+# SPDX-License-Identifier: GPL-2.0-or-later
+# Copyright 2022, AMD
+
 import requests
 import weakref
 from dataclasses import dataclass, field
@@ -25,7 +14,7 @@ from concurrent import futures
 import bpy.utils.previews
 
 from ..utils import logging, update_ui, MATLIB_DIR
-log = logging.Log('utils.matlib')
+log = logging.Log('matlib.manager')
 
 URL = "https://api.matlib.gpuopen.com/api"
