+++ materialx/matlib/properties.py
@@ -1,25 +1,10 @@
+# SPDX-License-Identifier: GPL-2.0-or-later
+# Copyright 2022, AMD
 
-#**********************************************************************
-# Copyright 2020 Advanced Micro Devices, Inc
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#********************************************************************
 import bpy
 
 from ..matlib.manager import manager
-from . import MaterialXProperties
-
-from ..utils import logging
-log = logging.Log(tag='properties.matlib')
+from ..utils import MaterialXProperties
 
 
 class MatlibProperties(bpy.types.PropertyGroup):
