+++ materialx/matlib/ui.py
@@ -1,17 +1,6 @@
-# **********************************************************************
-# Copyright 2020 Advanced Micro Devices, Inc
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#********************************************************************
+# SPDX-License-Identifier: GPL-2.0-or-later
+# Copyright 2022, AMD
+
 import traceback
 import textwrap
 
@@ -26,7 +15,7 @@ from .manager import manager
 #from .. import config
 
 from ..utils import logging
-log = logging.Log('ui.matlib')
+log = logging.Log('matlib.ui')
 
 
 class MATERIAL_OP_matlib_clear_search(bpy.types.Operator):
